requirements.txt
pandas
requests
psycopg2-binary
mailslurp_client